Ensure install_gfortran.sh preserves symlinks
When libquadmath and libgfortran libraries installed their symbolic
links are not preserved. Instead they are installed as duplicate files.

This causes ldconfig errors:

ldconfig: /app/lib/libquadmath.so.0 is not a symbolic link
ldconfig: /app/lib/libgfortran.so.3 is not a symbolic link

It still works, but it adds some confusion when less experienced users
file bug reports. Preserving symbolic links solves this.